+++ src/Components/Applets/AllApplets.vue
