+++ src/Components/CalendarComponents/ActivitySidebar.vue
