+++ src/Steps/SetSchedule.vue
@@ -12,8 +12,8 @@
 
 <script>
 import _ from 'lodash';
-import Calendar from '../Custom/CalendarComponents/CalendarMain';
-import adminApi from '../Custom/Utils/api';
+import Calendar from '../Components/CalendarComponents/CalendarMain';
+import adminApi from '../Components/Utils/api';
 
 export default {
   name: 'Schedule',
